# Changes to invariant based on comments