Ok without losing too much performance on `getCoord()` I added a number validation function at the end and checks for `null` & `false`.

~Added a param for the validation, that way you can turn it on or off. (default=false)~

**Benchmark before**

```
getCoord.Point x 18,555,451 ops/sec ±1.18% (86 runs sampled)
```

**Benchmark with number Validation**

```
getCoord.Point x 45,853,256 ops/sec ±0.63% (88 runs sampled)
getCoord.LineString x 33,147,142 ops/sec ±1.29% (91 runs sampled)
getCoord.Polygon x 27,853,174 ops/sec ±0.78% (92 runs sampled)
```

**Benchmark without number Validation**

```
getCoord.Point x 51,871,316 ops/sec ±1.49% (89 runs sampled)
getCoord.LineString x 49,007,217 ops/sec ±2.44% (86 runs sampled)
getCoord.Polygon x 52,104,843 ops/sec ±0.74% (88 runs sampled)
```

CC: @tmcw