@tmcw 2017-03-07
If we make validation optional, this change will have to be a major version bump for Turf. I'm kind of unconvinced here - like something that's hitting millions of ops / sec is taking millionths of seconds - like 21 nanoseconds. Shaving several nanoseconds in exchange for opening the door to uncaught errors doesn't seem like a worthwhile tradeoff.