@DenisCarriere 2017-03-07
I totally agree with your last comment, I've already removed the `validation` option.

We shouldn't add it as an option even for a major release.