@DenisCarriere 2017-03-07
I don't have any more commits to fire off, @tmcw I'll let you merge or let me know if you have any more comments and I can change them.

So far... nothing has changed the `@turf/invariant` behavior (still backwards compatible) except for the additional geometry support for `getCoord()`.