@tmcw 2017-03-07
I feel like this is going to break things: getCoord is singular, and the Point-geometry checking was intentional. The new code will let `getCoord(Polygon) -> Array<Array<number>>`, which will break getCoord's contract.

In current master:

```js
> var distance = require('./packages/turf-distance/');
undefined
> distance({ type: 'LineString', coordinates: [[0,0],[1,1]] }, [3,4]);
NaN
```

Before the geometry check was introduced:

```js
> var distance = require('./packages/turf-distance/');
undefined
> distance({ type: 'LineString', coordinates: [[0,0],[1,1]] }, [3,4]);
Error: A coordinate, feature, or point geometry is required
    at getCoord (/Users/tmcw/src/turf/packages/turf-invariant/index.js:24:11)
    at module.exports (/Users/tmcw/src/turf/packages/turf-distance/index.js:49:24)
```