@DenisCarriere 2017-03-07
**Proposed Solution**

Have two methods:

- `getCoord()` - Remains identical as before since it's meant to retrieve a single Coordinate (Point). Will throw an error if coordinates Array is not `Array<number>`.
- `getCoords()` - **New method** - retrieves coordinates which would normally be (Point/Line/Polygon) Also checks if array contains a number.

Thoughts?