@DenisCarriere 2017-03-08
@tmcw Last commit https://github.com/Turfjs/turf/pull/606/commits/bce258e5103d7d1a99dd7e5bf8d1ad18d9e6e91e should address all the concerns you mentioned above.

- Implement getCoords() which retrieves coordinates
- Make sure that getCoord() doesn't break any backwards compatibility
  - Test for Null
  - Test for False
  - Test for LineString
  - Test for Arrays that must be an array of two numbers
- Add getCoords() to typescript definition
- Update README